fix(material-experimental/mdc-slider): collapse everything into slider.ts
* delete slider-adapter.ts
* delete slider-interface.ts
* delete slider-thumb.ts
* move SliderAdapter implementation to slider.ts
* move MatSliderThumb implementation to slider.ts
* go back to using direct references instead of interfaces
* fix imports in module.ts
* fix imports in public-api.ts
