--- src/material-experimental/mdc-slider/slider-thumb.ts
@@ -1,144 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.io/license
- */
-
-import {coerceNumberProperty, NumberInput} from '@angular/cdk/coercion';
-import {DOCUMENT} from '@angular/common';
-import {
-  AfterViewInit,
-  Directive,
-  ElementRef,
-  EventEmitter,
-  Inject,
-  Input,
-  Output,
-} from '@angular/core';
-import {Thumb} from '@material/slider';
-import {_MatSliderInterface, MAT_SLIDER} from './slider-interface';
-
-/**
- * Represents a drag event emitted by the MatSlider component.
- */
-export interface MatSliderDragEvent {
-  /** The MatSliderThumb that was interacted with. */
-  source: MatSliderThumb;
-
-  /** The current value of the slider. */
-  value: number;
-
-  /** The thumb that was interacted with. */
-  thumb: Thumb;
-}
-
-/**
- * The native input used by the MatSlider.
- */
-@Directive({
-  selector: 'input[matSliderThumb], input[matSliderStartThumb], input[matSliderEndThumb]',
-  host: {
-    'class': 'mdc-slider__input',
-    'type': 'range',
-    '(blur)': '_blur.emit()',
-    '(focus)': '_focus.emit()',
-  },
-})
-export class MatSliderThumb implements AfterViewInit {
-
-  // ** IMPORTANT NOTE **
-  //
-  // The way `value` is implemented for MatSliderThumb goes against our standard practice. Normally
-  // we would define a private variable `_value` as the source of truth for the value of the slider
-  // thumb input. The source of truth for the value of the slider inputs has already been decided
-  // for us by MDC to be the value attribute on the slider thumb inputs. This is because the MDC
-  // foundation and adapter expect that the value attribute is the source of truth for the slider
-  // inputs.
-  //
-  // Also, note that the value attribute is completely disconnected from the value property.
-
-  /** The current value of this slider input. */
-  @Input()
-  get value(): number {
-    return coerceNumberProperty(this._elementRef.nativeElement.getAttribute('value'));
-  }
-  set value(v: number) {
-    const value = coerceNumberProperty(v);
-
-    // If the foundation has already been initialized, we need to
-    // relay any value updates to it so that it can update the UI.
-    if (this._slider._initialized) {
-      this._slider._setValue(value, this.thumb);
-    } else {
-      // Setup for the MDC foundation.
-      this._elementRef.nativeElement.setAttribute('value', `${value}`);
-    }
-  }
-
-  /** Event emitted when the slider thumb starts being dragged. */
-  @Output() readonly dragStart: EventEmitter<MatSliderDragEvent>
-    = new EventEmitter<MatSliderDragEvent>();
-
-  /** Event emitted when the slider thumb stops being dragged. */
-  @Output() readonly dragEnd: EventEmitter<MatSliderDragEvent>
-    = new EventEmitter<MatSliderDragEvent>();
-
-  /** Event emitted every time the MatSliderThumb is blurred. */
-  @Output() readonly _blur: EventEmitter<void> = new EventEmitter<void>();
-
-  /** Event emitted every time the MatSliderThumb is focused. */
-  @Output() readonly _focus: EventEmitter<void> = new EventEmitter<void>();
-
-  /** Indicates which slider thumb this input corresponds to. */
-  thumb: Thumb;
-
-  private _document: Document;
-
-  constructor(
-    @Inject(DOCUMENT) document: any,
-    @Inject(MAT_SLIDER) private readonly _slider: _MatSliderInterface,
-    readonly _elementRef: ElementRef<HTMLInputElement>,
-    ) {
-      this._document = document;
-      this.thumb = _elementRef.nativeElement.hasAttribute('matSliderStartThumb')
-        ? Thumb.START
-        : Thumb.END;
-
-      // Only set the default value if an initial value has not already been provided.
-      // Note that we are only setting the value attribute at this point. We cannot set the value
-      // property yet because the min and max have not been set.
-      if (!_elementRef.nativeElement.hasAttribute('value')) {
-        this.value = _elementRef.nativeElement.hasAttribute('matSliderEndThumb')
-          ? _slider.max
-          : _slider.min;
-      }
-    }
-
-  ngAfterViewInit() {
-    const min = this._elementRef.nativeElement.hasAttribute('matSliderEndThumb')
-      ? this._slider._getInput(Thumb.START).value
-      : this._slider.min;
-    const max = this._elementRef.nativeElement.hasAttribute('matSliderStartThumb')
-      ? this._slider._getInput(Thumb.END).value
-      : this._slider.max;
-    this._elementRef.nativeElement.min = `${min}`;
-    this._elementRef.nativeElement.max = `${max}`;
-
-    // We can now set the property value because the min and max have now been set.
-    this._elementRef.nativeElement.value = `${this.value}`;
-
-    // Setup for the MDC foundation.
-    if (this._slider.disabled) {
-      this._elementRef.nativeElement.disabled = true;
-    }
-  }
-
-  /** Returns true if this slider input currently has focus. */
-  _isFocused(): boolean {
-    return this._document.activeElement === this._elementRef.nativeElement;
-  }
-
-  static ngAcceptInputType_value: NumberInput;
-}
